extract `can_target` to `.permissions` module
