+++ pycroft/lib/user/edit.py
@@ -7,7 +7,8 @@ from pycroft.lib.logging import log_user_event
 from pycroft.model.session import with_transaction
 from pycroft.model.user import User
 
-from ._old import can_target, send_confirmation_email
+from ._old import send_confirmation_email
+from .permission import can_target
 
 
 @with_transaction
