+++ pycroft/lib/user/passwords.py
@@ -4,8 +4,7 @@ from pycroft.lib.logging import log_user_event
 from pycroft.model.session import with_transaction
 from pycroft.model.user import User
 
-from ._old import can_target
-
+from .permission import can_target
 
 
 def maybe_setup_wifi(user: User, processor: User) -> str | None:
